+++ web-common/src/features/dashboards/granular-access-policies/useSecurityPolicyCheck.ts
